+++ src/main/kotlin/net/nuclearprometheus/tpm/applicationserver/domain/ports/services/file/FileServiceImpl.kt
@@ -1,21 +1,22 @@
 package net.nuclearprometheus.tpm.applicationserver.domain.ports.services.file
 
 import net.nuclearprometheus.tpm.applicationserver.domain.exceptions.common.NotFoundException
+import net.nuclearprometheus.tpm.applicationserver.domain.exceptions.file.FileAccessException
+import net.nuclearprometheus.tpm.applicationserver.domain.exceptions.project.ProjectAccessException
 import net.nuclearprometheus.tpm.applicationserver.domain.model.file.File
 import net.nuclearprometheus.tpm.applicationserver.domain.model.file.FileId
 import net.nuclearprometheus.tpm.applicationserver.domain.model.project.ProjectId
-import net.nuclearprometheus.tpm.applicationserver.domain.model.user.UserId
+import net.nuclearprometheus.tpm.applicationserver.domain.model.project.ProjectRole
+import net.nuclearprometheus.tpm.applicationserver.domain.model.user.UserRole
 import net.nuclearprometheus.tpm.applicationserver.domain.ports.repositories.file.FileRepository
 import net.nuclearprometheus.tpm.applicationserver.domain.ports.repositories.project.ProjectRepository
-import net.nuclearprometheus.tpm.applicationserver.domain.ports.repositories.project.TeamMemberRepository
-import net.nuclearprometheus.tpm.applicationserver.domain.ports.repositories.user.UserRepository
 import net.nuclearprometheus.tpm.applicationserver.domain.ports.services.logging.Logger
+import net.nuclearprometheus.tpm.applicationserver.domain.ports.services.user.UserContextProvider
 
 class FileServiceImpl(
     private val fileRepository: FileRepository,
-    private val teamMemberRepository: TeamMemberRepository,
-    private val userRepository: UserRepository,
     private val projectRepository: ProjectRepository,
+    private val userContextProvider: UserContextProvider,
     private val logger: Logger
 ) : FileService {
 
@@ -23,22 +24,22 @@ class FileServiceImpl(
         name: String,
         size: Long,
         type: String,
-        uploaderId: UserId,
         projectId: ProjectId,
         location: String
     ): File {
-        projectRepository.get(projectId) ?: throw NotFoundException("Project with id $projectId does not exist")
+        val currentUser = userContextProvider.getCurrentUser()
+        val project = projectRepository.get(projectId) ?: throw NotFoundException("Project with id $projectId does not exist")
 
-        val uploader = userRepository.get(uploaderId)
-            ?: throw NotFoundException("User with id $uploaderId does not exist")
-        teamMemberRepository.getByUserIdAndProjectId(uploaderId, projectId)
-            ?: throw NotFoundException("User with id $uploaderId is not a member of project with id $projectId")
+        if (!currentUser.hasRole(UserRole.ADMIN) && !project.hasTeamMember(currentUser.id)) {
+            logger.error("User ${currentUser.id} is not a member of project $projectId")
+            throw ProjectAccessException("User ${currentUser.id} is not a member of project $projectId")
+        }
 
         val file = File(
             name = name,
             size = size,
             type = type,
-            uploader = uploader,
+            uploader = currentUser,
             projectId = projectId,
             location = location
         )
@@ -47,6 +48,20 @@ class FileServiceImpl(
     }
 
     override fun delete(id: FileId) {
+        val currentUser = userContextProvider.getCurrentUser()
+        val file = fileRepository.get(id) ?: throw NotFoundException("File with id $id does not exist")
+        val project = projectRepository.get(file.projectId) ?: throw IllegalStateException("Project with id ${file.projectId} does not exist")
+
+        if (!currentUser.hasRole(UserRole.ADMIN) && !project.hasTeamMember(currentUser.id)) {
+            logger.error("User ${currentUser.id} is not a member of project ${file.projectId}")
+            throw ProjectAccessException("User ${currentUser.id} is not a member of project ${file.projectId}")
+        }
+
+        if (file.uploader.id != currentUser.id && !currentUser.hasRole(UserRole.ADMIN) && !project.hasTeamMemberWithRole(currentUser.id, ProjectRole.PROJECT_MANAGER)) {
+            logger.error("User ${currentUser.id} is not the uploader of file $id and is not a project manager")
+            throw FileAccessException("User ${currentUser.id} is not the uploader of file $id and is not a project manager")
+        }
+
         fileRepository.delete(id)
     }
 }
